Ya pochinilgit statusgit statusgit statusgit statusgit statusgit statusgit statusgit status

+++ nodeActions.cpp
@@ -1,23 +1,23 @@
 #include "nodeActions.h"
 
-double rotateNodeX(const int xCenter, const int secondCenter,
-                 const int xCoord, const int secondCoord,
+double rotateNodeX(const double xCenter, const double secondCenter,
+                 const double xCoord, const double secondCoord,
                  const double cosAngle, const double sinAngle) {
     return xCenter +
             (xCoord - xCenter) * cosAngle +
             (secondCoord - secondCenter) * sinAngle;
 }
 
-double rotateNodeY(const int yCenter, const int secondCenter,
-                 const int yCoord, const int secondCoord,
+double rotateNodeY(const double yCenter, const double secondCenter,
+                 const double yCoord, const double secondCoord,
                  const double cosAngle, const double sinAngle) {
     return yCenter +
             (yCoord - yCenter) * cosAngle -
             (secondCoord - secondCenter) * sinAngle;
 }
 
-double rotateNodeZ(const int zCenter, const int secondCenter,
-                 const int zCoord, const int secondCoord,
+double rotateNodeZ(const double zCenter, const double secondCenter,
+                 const double zCoord, const double secondCoord,
                  const double cosAngle, const double sinAngle) {
     return zCenter +
             (zCoord - zCenter) * cosAngle -
